Add a statline to an scplot

Add a statline to an scplot

add_statline(
  object,
  stat,
  phase = NULL,
  color = NULL,
  linewidth = NULL,
  size = NULL,
  linetype = NULL,
  variable = NULL,
  ...
)

Arguments

object

An scplot object (class `scplot`) returned from the `scplot()` function.

stat

A character string for defining a statistical line or curve to be plotted. Possible values: "median", "mean", "min", "max", "quantile", "trend", "trendA", "trendA_bisplit", "trendA_trisplit", "movingMean", "movingMedian", "loreg".

phase

Either a numeric or a character vector specifying the reference phase (see details)

color

A character string or a number defining the color of an element.

linewidth

A number with the width of the line.

size

Text size relative to the base text size.

linetype

A character string with the line type: "solid", "dashed", "dotted"

variable

Name of the dataline variable to apply the style.

...

additional parameters passed to the statistical function.

Details

The 'phase' argument defines the reference phase for some statistiscal functions ("median", "mean", "min", "max", "quantile"). The default is NULL which calculates and plots statistics for each phase separately. The arguments takes a numeric vector (phase number(s)) or a character vector (phase name(s)). When more than one phase is defines, statistics are based on the combined values of these phases. Some of the functions defined in 'stats' have additional arguments. The 'mean' function has a trim argument (e.g. 'trim = 0.1'). 'quantile has a proportion argument (e.g. prob = 0.75 for calculating the 75 'trendA' has the argument 'method' which takes 'ols' or 'theil-sen'. 'movingMean' and 'movingMedian' have a lag argument (e.g. lag = 2). The local-regression curve function 'lowess' (or 'loreg') has a proportion argument (e.g. f = 0.5) and the local-regression curve function 'loess' has a span argument (e.g. span = 0.75).