Plot single-case data

This function provides a plot of a single-case or multiple single-cases.

scplot(data)

set_xlabel(
  object,
  label = NULL,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL
)

set_ylabel(
  object,
  label,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL
)

set_xaxis(
  object,
  limits = NULL,
  increment = NULL,
  increment_from = NULL,
  color = NULL,
  size = NULL,
  line = NULL,
  angle = NULL,
  positions = NULL
)

set_yaxis(
  object,
  limits = NULL,
  color = NULL,
  size = NULL,
  increment = NULL,
  increment_from = NULL,
  positions = NULL
)

add_title(
  object,
  label,
  color = NULL,
  size = NULL,
  family = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  margin = NULL,
  parse = NULL
)

add_caption(
  object,
  label,
  color = NULL,
  size = NULL,
  face = NULL,
  hjust = NULL,
  vjust = NULL,
  wrap = NULL,
  margin = NULL,
  parse = NULL
)

set_casenames(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  width = NULL,
  angle = NULL,
  fill = NULL,
  linetype = NULL,
  frame = NULL,
  type = NULL
)

set_phasenames(
  object,
  ...,
  color = NULL,
  size = NULL,
  x = NULL,
  y = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  box = NULL,
  frame = NULL,
  face = NULL,
  family = NULL
)

set_seperator(
  object,
  color = NULL,
  width = NULL,
  linetype = NULL,
  extent = NULL,
  label = NULL,
  size = NULL
)

add_grid(
  object,
  color = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  arrow = NULL
)

add_labels(
  object,
  color = NULL,
  size = NULL,
  nudge_y,
  nudge_x,
  vjust = NULL,
  hjust = NULL,
  angle = NULL,
  round,
  box = NULL,
  frame = NULL
)

add_dataline(object, variable, color, width, linetype, dots, shape, size)

set_dataline(object, variable, color, width, linetype, dots, shape, size)

add_statline(
  object,
  stat,
  color = NULL,
  width = NULL,
  linetype = NULL,
  variable = NULL,
  ...
)

set_background(object, color = NULL, fill = NULL, linetype = NULL, size = NULL)

set_panel(
  object,
  color = NULL,
  fill = NULL,
  linetype = NULL,
  size = NULL,
  alpha = NULL
)

add_theme(object, theme, ...)

add_marks(
  object,
  case = 1,
  positions,
  color = "red",
  size = 1,
  shape = 1,
  variable = ".dvar"
)

add_text(object, label, case = 1, x, y, color = "black", size = 2, angle = 0)

add_arrow(
  object,
  case = 1,
  x0,
  y0,
  x1,
  y1,
  color = "black",
  angle = 30,
  length = unit(5, "points"),
  ends = "last",
  type = "open"
)

set_theme_element(object, ...)

add_ridge(object, color = "grey98", variable = ".dvar")

add_legend(
  object,
  labels = ".default",
  case = 1,
  position = "right",
  datalines = TRUE,
  statlines = TRUE,
  title = NULL
)

Arguments

label
Character string.

color
A character string or a number defining the color of an element.

size
A number deifning the size of an element.

limits
Lower and upper limits of the axis (e.g., limits = c(0, 20) sets the axis to a scale from 0 to 20). With multiple single-cases you can use limits = c(0, NA) to scale the axis from 0 to the maximum of each case. limits is not set by default, which makes scan set a proper scale based on the given data.

increment
An integer. Increment of the x-axis. 1 :each mt value will be printed, 2 : every other value, 3 : every third values etc.

positions
Either a vector indicating the dots to be highlighted or a character string with a logical expression (e.g. values < mean(values))

parse
If TRUE, the label is interpreted as an expression. Default = FALSE.

wrap
Number that defines the maximum characters per line before a break. If set to FALSE, no automatic linebreak is set.

...
various style parameter

x
x position

y
y position

width
A number defining the line width

type
A character string with the line type: "solid", "dashed", "dotted"

extent
A number between 0 and 1 given the proportion of the plot that is covert by the line or character string "full" or "scale".

round
Number of digits of the labels.

shape
Number. See pch graphical parameter on par help page.

stat
A character string defining a line or curve to be plotted. Possible values: "median", "mean", "trend", "trendA", "trendA_bisplit", "trendA_trisplit", "maxA", "minA", "meanA", "medianA", "plm", "movingMean", "movingMedian", "loreg"

theme
A character string with a predefined graphical theme. Possible values: default, yaxis, tiny, small, big, chart, ridge, annotate, grid, grid2, dark, nodots, sienna, phase_color, phase_shade

case
Numerical vector with the csae number or character string "all" for all cases.

x0
Origin x position of the line.

y0
Origin y position of the line.

x1
End x position of the line.

y1
End y position of the line.

length
Size of the aroow angels.

orientation
of the label: 0 = vertical; 1 = horizontal

align
Character string. One of "left", "right", "center"

Value

An scplot object that creates a ggplot2 plot when printed.

Author

Juergen Wilbert