Add arrrows to an scplot

Add arrrows to an scplot

add_arrow(
  object,
  case = 1,
  x0,
  y0,
  x1,
  y1,
  color = "black",
  angle = 30,
  length = unit(5, "points"),
  type = "open",
  ends = "last",
  linewidth = 0.7
)

Arguments

object

An scplot object (class scplot) returned from the scplot() function.

case

Numerical vector with the case number or character string. case = "all" for all cases.

x0

Origin x position of the line.

y0

Origin y position of the line.

x1

End x position of the line.

y1

End y position of the line.

color

A character string or a number defining the color of an element.

angle

Angle (in [0,360])

length

Size of the arrow angels.

type

One of "open" or "closed" indicating whether the arrow head should be a closed triangle.

ends

One of "last", "first", or "both", indicating which ends of the line to draw arrow heads.

linewidth

A number with the width of the line.

Examples

p1 <- scplot(exampleAB$Anja)  |>
  add_arrow(case = 1, 2, 70, 6, 55, color = "darkred")